# docs/specification/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "description": "Versioned API for sports venue business management clients and integrations.",
    "title": "Sports Venue Management API",
    "version": "0.1.0"
  },
...
  "paths": {
...
    "/v1/activities/{activityId}": {
      "patch": {
        "operationId": "updateActivity",
        "tags": [
          "configuration"
        ],
        "description": "Edit an activity using optimistic concurrency.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "anyOf": [
                  {
                    "required": [
                      "displayName"
                    ]
                  },
                  {
                    "required": [
                      "state"
                    ]
                  }
                ],
                "properties": {
                  "displayName": {
                    "maxLength": 120,
                    "minLength": 1,
                    "type": "string"
                  },
                  "expectedVersion": {
                    "minimum": 1,
                    "type": "integer"
                  },
                  "state": {
                    "enum": [
                      "ACTIVE",
                      "INACTIVE"
                    ],
                    "type": "string"
                  }
                },
                "required": [
                  "expectedVersion"
                ],
                "type": "object"
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "format": "uuid",
              "type": "string"
            },
            "in": "path",
            "name": "activityId",
            "required": true
          },
          {
            "schema": {
              "format": "uuid",
              "type": "string"
            },
            "in": "header",
            "name": "x-business-id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "activity": {
                      "additionalProperties": false,
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "createdAt": {
                          "format": "date-time",
                          "type": "string"
                        },
                        "displayName": {
                          "type": "string"
                        },
                        "id": {
                          "format": "uuid",
                          "type": "string"
                        },
                        "state": {
                          "enum": [
                            "ACTIVE",
                            "INACTIVE"
                          ],
                          "type": "string"
                        },
                        "updatedAt": {
                          "format": "date-time",
                          "type": "string"
                        },
                        "version": {
                          "minimum": 1,
                          "type": "integer"
                        }
                      },
                      "required": [
                        "code",
                        "createdAt",
                        "displayName",
                        "id",
                        "state",
                        "updatedAt",
                        "version"
                      ],
                      "type": "object"
                    },
                    "requestId": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "activity",
                    "requestId"
                  ],
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "requestId": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "requestId"
                  ],
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "requestId": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "requestId"
                  ],
                  "type": "object"
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "requestId": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "requestId"
                  ],
                  "type": "object"
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "requestId": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "requestId"
                  ],
                  "type": "object"
                }
              }
            }
          },
          "409": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "requestId": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "requestId"
                  ],
                  "type": "object"
                }
              }
            }
          },
          "422": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "requestId": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "requestId"
                  ],
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:4000"
    }
  ]
}